feat: support --sanity_check option, fix #6; and better error msg when checker failed

+++ src/witness_generator.ts
@@ -4,10 +4,10 @@ import * as shelljs from 'shelljs';
 const si = require('systeminformation');
 const { stringifyBigInts, unstringifyBigInts } = require('ffjavascript').utils;
 const { wtns } = require('snarkjs');
+import { groupOrderPrimeStr, groupOrderPrime } from './math';
 //import {compiler} from "circom";
 //const Scalar = require("ffjavascript").Scalar;
 
-const primeStr = '21888242871839275222246405745257275088548364400416034343698204186575808495617';
 const NODE_CMD = 'NODE_OPTIONS=--max-old-space-size=8192 node --stack-size=65500';
 
 function shellExecFnBuilder(verbose) {
@@ -28,19 +28,26 @@ async function compileWasmBinary({ circuitDirName, r1csFilepath, circuitFilePath
   shellExec(cmd);
 }
 
-async function compileNativeBinary({ circuitDirName, r1csFilepath, circuitFilePath, symFilepath, binaryFilePath, verbose }) {
+async function generateSrcsForNativeBinary({ circuitDirName, r1csFilepath, circuitFilePath, symFilepath, verbose, alwaysRecompile }) {
   const shellExec = shellExecFnBuilder(verbose);
   const circomRuntimePath = path.join(require.resolve('circom_runtime'), '..', '..');
   const ffiasmPath = path.join(require.resolve('ffiasm'), '..');
   const circomcliPath = path.join(require.resolve('circom'), '..', 'cli.js');
   const cFilepath = path.join(circuitDirName, 'circuit.cpp');
 
+  if (!alwaysRecompile && fs.existsSync(cFilepath)) {
+    if (verbose) {
+      console.log('skip generate c src', cFilepath);
+    }
+    return;
+  }
+
   let cmd: string;
   cmd = `cp ${circomRuntimePath}/c/*.cpp ${circuitDirName}`;
   shellExec(cmd);
   cmd = `cp ${circomRuntimePath}/c/*.hpp ${circuitDirName}`;
   shellExec(cmd);
-  cmd = `node ${ffiasmPath}/src/buildzqfield.js -q ${primeStr} -n Fr`;
+  cmd = `node ${ffiasmPath}/src/buildzqfield.js -q ${groupOrderPrimeStr} -n Fr`;
   shellExec(cmd, { cwd: circuitDirName });
   if (process.arch !== 'x64') {
     throw 'Unsupported platform ' + process.arch + '. Try wasm backend as an alternative';
@@ -51,24 +58,53 @@ async function compileNativeBinary({ circuitDirName, r1csFilepath, circuitFilePa
     cmd = `nasm -felf64 ${circuitDirName}/fr.asm`;
   } else throw 'Unsupported platform';
   shellExec(cmd);
+
   cmd = `${NODE_CMD} ${circomcliPath} ${circuitFilePath} -r ${r1csFilepath} -c ${cFilepath} -s ${symFilepath}`;
   shellExec(cmd);
-  if (process.platform === 'darwin') {
-    cmd = `g++ ${circuitDirName}/main.cpp ${circuitDirName}/calcwit.cpp ${circuitDirName}/utils.cpp ${circuitDirName}/fr.cpp ${circuitDirName}/fr.o ${cFilepath} -o ${binaryFilePath} -lgmp -std=c++11 -O3 -DSANITY_CHECK`;
-  } else if (process.platform === 'linux') {
-    cmd = `g++ -pthread ${circuitDirName}/main.cpp ${circuitDirName}/calcwit.cpp ${circuitDirName}/utils.cpp ${circuitDirName}/fr.cpp ${circuitDirName}/fr.o ${cFilepath} -o ${binaryFilePath} -lgmp -std=c++11 -O3 -fopenmp -DSANITY_CHECK`;
-  } else throw 'Unsupported platform';
-  shellExec(cmd);
+
+  // the binary needs a $arg0.dat file, so we make a symbol link here
+  cmd = `ln -s circuit.dat circuit.fast.dat`;
+  shellExec(cmd, { cwd: circuitDirName });
 }
 
-async function compileCircuitDir(circuitDirName, { alwaysRecompile, verbose, backend }) {
+async function compileNativeBinary({
+  circuitDirName,
+  r1csFilepath,
+  circuitFilePath,
+  symFilepath,
+  binaryFilePath,
+  verbose,
+  sanityCheck,
+  alwaysRecompile,
+}) {
+  await generateSrcsForNativeBinary({ circuitDirName, r1csFilepath, circuitFilePath, symFilepath, verbose, alwaysRecompile });
+  const shellExec = shellExecFnBuilder(verbose);
+  let compileCmd = `g++ ${circuitDirName}/main.cpp ${circuitDirName}/calcwit.cpp ${circuitDirName}/utils.cpp ${circuitDirName}/fr.cpp ${circuitDirName}/fr.o ${circuitDirName}/circuit.cpp -o ${binaryFilePath} -lgmp -std=c++11 -O3`;
+  if (process.platform === 'darwin') {
+    // do nothing
+  } else if (process.platform === 'linux') {
+    compileCmd += ' -pthread -fopenmp';
+  } else {
+    throw 'Unsupported platform';
+  }
+  if (sanityCheck) {
+    compileCmd += ' -DSANITY_CHECK';
+  }
+  shellExec(compileCmd);
+}
+
+async function compileCircuitDir(circuitDirName, { alwaysRecompile, verbose, backend, sanityCheck }) {
   // console.log('compiling dir', circuitDirName);
   const circuitFilePath = path.join(circuitDirName, 'circuit.circom');
   const r1csFilepath = path.join(circuitDirName, 'circuit.r1cs');
   const symFilepath = path.join(circuitDirName, 'circuit.sym');
   let binaryFilePath: string;
   if (backend === 'native') {
-    binaryFilePath = path.join(circuitDirName, 'circuit');
+    if (sanityCheck) {
+      binaryFilePath = path.join(circuitDirName, 'circuit');
+    } else {
+      binaryFilePath = path.join(circuitDirName, 'circuit.fast');
+    }
   } else {
     binaryFilePath = path.join(circuitDirName, 'circuit.wasm');
   }
@@ -81,8 +117,18 @@ async function compileCircuitDir(circuitDirName, { alwaysRecompile, verbose, bac
   console.log('compile', circuitDirName);
 
   if (backend === 'native') {
-    await compileNativeBinary({ circuitDirName, r1csFilepath, circuitFilePath, symFilepath, binaryFilePath, verbose });
+    await compileNativeBinary({
+      circuitDirName,
+      r1csFilepath,
+      circuitFilePath,
+      symFilepath,
+      binaryFilePath,
+      verbose,
+      sanityCheck,
+      alwaysRecompile,
+    });
   } else {
+    // sanity check is not supported for wasm backend now
     await compileWasmBinary({ circuitDirName, r1csFilepath, circuitFilePath, symFilepath, binaryFilePath, verbose });
   }
   return { circuitFilePath, r1csFilepath, symFilepath, binaryFilePath };
@@ -95,15 +141,25 @@ class WitnessGenerator {
   circuitDirName: string;
   binaryFilePath: string;
   alwaysRecompile: boolean;
+  sanityCheck: boolean;
   verbose: boolean;
   writeExpectedOutput: boolean;
   backend: string;
-  constructor(name, { backend, alwaysRecompile, verbose } = { backend: 'native', alwaysRecompile: true, verbose: false }) {
+  constructor(
+    name,
+    { backend, alwaysRecompile, verbose, sanityCheck } = { backend: 'native', alwaysRecompile: true, verbose: false, sanityCheck: true },
+  ) {
     this.name = name;
     // we can specify cached files to avoid compiling every time
     this.alwaysRecompile = alwaysRecompile;
     this.verbose = verbose;
     this.backend = backend;
+    this.sanityCheck = sanityCheck;
+    if (this.verbose) {
+      if (this.backend == 'wasm' && this.sanityCheck) {
+        console.log('WARN: sanity check is not supported for wasm backend now');
+      }
+    }
   }
 
   async compile(circuitDirName) {
@@ -125,6 +181,7 @@ class WitnessGenerator {
       alwaysRecompile: this.alwaysRecompile,
       verbose: this.verbose,
       backend: this.backend,
+      sanityCheck: this.sanityCheck,
     });
     this.binaryFilePath = binaryFilePath;
     return { r1csFilepath, symFilepath };
@@ -141,10 +198,16 @@ class WitnessGenerator {
     // gen witness
     if (this.backend === 'native') {
       cmd = `${this.binaryFilePath} ${inputFilePath} ${witnessFilePath}`;
-      const genWtnsOut = shelljs.exec(cmd);
+      if (this.verbose) {
+        console.log(cmd);
+      }
+      const genWtnsOut = shelljs.exec(cmd, { silent: !this.verbose });
 
       if (genWtnsOut.stderr || genWtnsOut.code != 0) {
-        console.error(genWtnsOut.stderr);
+        if (!this.verbose) {
+          // don't display stderr twice
+          console.error('\n' + genWtnsOut.stderr);
+        }
         throw new Error('Could not generate witness');
       }
     } else {
